+++ src/components/FuncButtons/DeleteCard/index.jsx
