+++ src/components/FuncButtons/Styles/CheckBoxStyled.jsx
